requirements.txt
flask>=3.0.0
websocket-client>=1.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.0.47
nacos-sdk-python>=2.0.0,<3.0.0
chromadb>=0.5.0
FlagEmbedding>=1.2.0
